New Podiums for Standings
Recreated the podiums. I didn't like the way they looked. Added a linear gradient, made them smaller and separated them and added border-radius of 5px. Took the names and scoring out of the podium block and moved them to a footer underneath. Changed the color of the avatar border to match the main podium color.